# specs/api/openapi.yaml
openapi: 3.1.0
info:
  title: UniChat API
  description: UniChat unified communication platform API
  version: 1.0.0
paths:
...
  /workspaces:
    get:
      tags:
      - workspaces_and_channels
      summary: List workspaces for current user
      operationId: list_workspaces_workspaces_get
      parameters:
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceResponse'
                title: Response List Workspaces Workspaces Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workspaces_and_channels
      summary: Create a new workspace
      operationId: create_workspace_workspaces_post
      parameters:
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}:
    delete:
      tags:
      - workspaces_and_channels
      summary: Delete a workspace (Owner only)
      operationId: delete_workspace_workspaces__workspace_id__delete
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Delete Workspace Workspaces  Workspace Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/leave:
    post:
      tags:
      - workspaces_and_channels
      summary: Leave a workspace (removes from workspace and all channels; owner cannot
        leave)
      operationId: leave_workspace_workspaces__workspace_id__leave_post
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Leave Workspace Workspaces  Workspace Id  Leave Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/members/{user_id}:
    delete:
      tags:
      - workspaces_and_channels
      summary: Remove a member from a workspace (Owner only)
      operationId: remove_workspace_member_workspaces__workspace_id__members__user_id__delete
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: user_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: User Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Remove Workspace Member Workspaces  Workspace Id  Members  User
                  Id  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/members:
    get:
      tags:
      - workspaces_and_channels
      summary: List workspace members
      operationId: list_workspace_members_workspaces__workspace_id__members_get
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WorkspaceMemberResponse'
                title: Response List Workspace Members Workspaces  Workspace Id  Members
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/invites:
    post:
      tags:
      - workspaces_and_channels
      summary: Create a workspace invite link
      operationId: create_invite_workspaces__workspace_id__invites_post
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateInviteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /invites/{token}:
    get:
      tags:
      - workspaces_and_channels
      summary: Preview workspace invite link details
      operationId: get_invite_preview_invites__token__get
      parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
          title: Token
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitePreviewResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /invites/{token}/accept:
    post:
      tags:
      - workspaces_and_channels
      summary: Accept workspace invite
      operationId: accept_invite_invites__token__accept_post
      parameters:
      - name: token
        in: path
        required: true
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInviteResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /workspaces/{workspace_id}/channels:
    get:
      tags:
      - workspaces_and_channels
      summary: List channels in workspace
      operationId: list_workspace_channels_workspaces__workspace_id__channels_get
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelResponse'
                title: Response List Workspace Channels Workspaces  Workspace Id  Channels
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - workspaces_and_channels
      summary: Create a channel in workspace
      operationId: create_channel_workspaces__workspace_id__channels_post
      parameters:
      - name: workspace_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Workspace Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChannelRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChannelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /channels/{channel_id}/join:
    post:
      tags:
      - workspaces_and_channels
      summary: Join a channel
      operationId: join_channel_channels__channel_id__join_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Join Channel Channels  Channel Id  Join Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /channels/{channel_id}/leave:
    post:
      tags:
      - workspaces_and_channels
      summary: Leave a channel
      operationId: leave_channel_channels__channel_id__leave_post
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Leave Channel Channels  Channel Id  Leave Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /channels/{channel_id}/members:
    get:
      tags:
      - workspaces_and_channels
      summary: List channel members
      operationId: list_channel_members_channels__channel_id__members_get
      parameters:
      - name: channel_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Channel Id
      - name: token
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Token
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChannelMemberResponse'
                title: Response List Channel Members Channels  Channel Id  Members
                  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    AcceptInviteResponse:
      properties:
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
      type: object
      required:
      - workspace_id
      title: AcceptInviteResponse
...
    ChannelMemberResponse:
      properties:
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        avatar_color:
          type: string
          title: Avatar Color
        joined_at:
          type: string
          format: date-time
          title: Joined At
      type: object
      required:
      - user_id
      - name
      - email
      - avatar_color
      - joined_at
      title: ChannelMemberResponse
    ChannelResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        name:
          type: string
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_by:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Created By
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - workspace_id
      - name
      - created_at
      title: ChannelResponse
...
    CreateChannelRequest:
      properties:
        name:
          type: string
          maxLength: 80
          minLength: 1
          title: Name
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
      type: object
      required:
      - name
      title: CreateChannelRequest
    CreateInviteResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        workspace_id:
          type: string
          format: uuid
          title: Workspace Id
        token:
          type: string
          title: Token
        expires_at:
          type: string
          format: date-time
          title: Expires At
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - workspace_id
      - token
      - expires_at
      - created_at
      title: CreateInviteResponse
...
    CreateWorkspaceRequest:
      properties:
        name:
          type: string
          maxLength: 100
          minLength: 1
          title: Name
      type: object
      required:
      - name
      title: CreateWorkspaceRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    InvitePreviewResponse:
      properties:
        workspace_name:
          type: string
          title: Workspace Name
        inviter_name:
          type: string
          title: Inviter Name
        expires_at:
          type: string
          format: date-time
          title: Expires At
        is_expired:
          type: boolean
          title: Is Expired
      type: object
      required:
      - workspace_name
      - inviter_name
      - expires_at
      - is_expired
      title: InvitePreviewResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    WorkspaceMemberResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        user_id:
          type: string
          format: uuid
          title: User Id
        name:
          type: string
          title: Name
        email:
          type: string
          title: Email
        avatar_color:
          type: string
          title: Avatar Color
        role:
          type: string
          title: Role
        joined_at:
          type: string
          format: date-time
          title: Joined At
        online:
          type: boolean
          title: Online
          default: false
      type: object
      required:
      - id
      - user_id
      - name
      - email
      - avatar_color
      - role
      - joined_at
      title: WorkspaceMemberResponse
    WorkspaceResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        name:
          type: string
          title: Name
        owner_id:
          type: string
          format: uuid
          title: Owner Id
        created_at:
          type: string
          format: date-time
          title: Created At
        member_count:
          type: integer
          title: Member Count
          default: 1
      type: object
      required:
      - id
      - name
      - owner_id
      - created_at
      title: WorkspaceResponse